social_network/requirements.txt
# requirements.txt

asgiref==3.8.1
Django==5.0.2
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.4.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
